Add the ability to use Asio without Boost.
When using a C++11 compiler, most of Asio may now be used without a
dependency on Boost header files or libraries. To use Asio in this
way, define ASIO_STANDALONE on your compiler command line or as part of
the project options.

The standalone configuration has currently been tested for the following
platforms and compilers:

- Linux with g++ 4.7 (requires -std=c++11)

- Mac OS X with clang++ / Xcode 4.6 (requires -std=c++11 -stdlib=libc++)